drop of support for 6.0
substantial simplification of code by dropping support for 6.0. We now have a single jar file, compiled to be compatible with java 8 and newer.

Also, shift to newer jpype functionality for importing NetLogoLink which should give more information if class cannot be found.

code is tested on m1 mac against 6.3 and against 6.1, 6.2 and 6.3 on windows